requirements.txt
asn1crypto==1.5.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cloudpickle==2.2.1
cryptography==43.0.0
filelock==3.15.4
idna==3.8
numpy==2.1.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests
requests-oauthlib
six==1.16.0
snowflake-connector-python
snowflake-snowpark-python
sortedcontainers==2.4.0
toml==0.10.2
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
